admin

+++ meu-app-laravel/routes/web.php
@@ -15,7 +15,7 @@ Route::get('/', function () {
 });
 
 Route::middleware(['auth'])->group(function () {
-    Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy');
+    Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy')->middleware('auth');
     Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
     Route::post('/user', [UserController::class, 'store'])->name('users.store');
     Route::get('/users', [UserController::class, 'index'])->name('users.index');
@@ -25,6 +25,10 @@ Route::middleware(['auth'])->group(function () {
 
     Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
     Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show');
+});
+
+Route::middleware(['auth', 'admin'])->group(function(){
+    Route::get('/admin', [UserController::class, 'admin'])->name('admin');
 });
